feat: Add DoneTasksInfo component to TaskBoard
This commit adds a new component called DoneTasksInfo to the TaskBoard. The DoneTasksInfo component displays information about completed tasks and is rendered in the TaskColumnContent when the status is 'done'. This component includes an icon and a message indicating that completed tasks are cleared at the start of each week.