feat: Add useUpdateUserEntityMutation hook for updating task status
This commit adds the useUpdateUserEntityMutation hook to the FocusProvider component. The hook is used to update the status of a task when a session is completed. The task status is set to 'inProgress' if the task is currently in progress. This ensures that the task status is correctly updated in the UI.