refactor: Remove unused date-fns imports and update date calculations
This commit removes the unused imports of `getWeekEndedAt` and `getWeekStartedAt` from the `lib/utils/time` directory. It also updates the date calculations in the `isThisWeek` and `isNextWeek` functions to use the `startOfISOWeek` and `endOfISOWeek` functions from the `date-fns` library. These changes improve the efficiency and readability of the code related to week calculations.