Feat/calendar (#72)
* feat: Update ExpandableSelector options in TaskDeadlineInput and TaskStatusInput

This commit updates the ExpandableSelector options in the TaskDeadlineInput and TaskStatusInput components. The changes include setting the floatingOptionsWidthSameAsOpener prop to false and the showToggle prop to false. These modifications improve the user experience by providing a more streamlined and intuitive interface for selecting task deadlines and statuses.

* feat: Update ExpandableSelector options in TaskDeadlineInput and TaskStatusInput

This commit updates the ExpandableSelector options in the TaskDeadlineInput and TaskStatusInput components. The changes include setting the floatingOptionsWidthSameAsOpener prop to false and the showToggle prop to false. These modifications improve the user experience by providing a more streamlined and intuitive interface for selecting task deadlines and statuses.

* feat: Improve CalendarDayInput and WeekdayLabels UI

This commit updates the CalendarDayInput and WeekdayLabels components to improve the user interface. Changes include adding a larger size to the navigation buttons in CalendarDayInput and adjusting the grid template rows in WeekdayLabels. These modifications enhance the visual appearance and usability of the calendar feature.

* refactor: Remove unused date-fns imports and update date calculations

This commit removes the unused imports of `getWeekEndedAt` and `getWeekStartedAt` from the `lib/utils/time` directory. It also updates the date calculations in the `isThisWeek` and `isNextWeek` functions to use the `startOfISOWeek` and `endOfISOWeek` functions from the `date-fns` library. These changes improve the efficiency and readability of the code related to week calculations.