feat: Improve TaskTrackedTime component
This commit adds a check in the TaskTrackedTime component to hide it if the spentTime is less than 1 minute. This check is implemented by using the convertDuration function from the @lib/utils/time/convertDuration module to convert the spentTime from milliseconds to minutes and rounding it to the nearest whole number. If the converted value is less than 1, the component is returned as null.

These changes enhance the usability of the TaskTrackedTime component by preventing it from being displayed when the spentTime is negligible, improving the overall user experience.