feat: Update FocusProvider component in FocusSetWidget
This commit adds an import statement for the 'shouldBePresent' function from the '@lib/utils/assert/shouldBePresent' module in the FocusProvider component. The 'shouldBePresent' function is used to ensure that the 'intervals' array in the 'session' object is present before accessing its last item. This change improves the reliability and stability of the FocusProvider component, enhancing the overall user experience in the FocusSetWidget.